Move color package to common
