+++ src/internal/common/color/color.go
